persist: remove WriteHandle::{append,append_batch}
The original inclusion of `append` in the persist API was targeted at
multi-replica clusters. The thinking was that each replica could assume
the other was writing the same data (definite inputs => definite
outputs) so if it was writing `[X, X+2)` and found that the upper of the
shard was `X=1`, it could simply truncate to `[X+1,X=2)` and try again.

To support this truncation in an efficient way (i.e. without writing it
back out to s3), persist had an internal optimization for append. We'd
link the same s3 data in, but with metadata indicating that irrelevant
data should be filtered out.

This internal performance optimization is the source of some
considerable complexity in persist. Furthermore, the imagined usage was
replaced by the self-correcting persist sink. The only callers of
`append` in production code were giving it an empty set of updates,
which is efficient _without_ the complex optimization.

So, remove it. Existing callers are replaced by manually unrolling the
CaA expectation-mismatch loop. I considered leaving an `append_empty` in
persist to DRY, but I've always found the semantics of `append` tricky
to explain and reason about and IMO the callers read more clearly with
the small bit of duplication.

This removes the API of `append`, but because of backward compatibility,
does not yet remove the truncation optimization. The code at HEAD
doesn't use append in any meaningful way, but it's possible that a
previous version did and that we have these sitting in prod. We'll have
to figure out a way to confirm their non-existence before we can do the
rest of the cleanup.