compute*: introduce `Schedule` command
This commit introduces a new `Schedule` command to the compute protocol.
The command is meant to delay the hydration of dataflows, as a means to
control ensure replica don't spend their hydration capacity on dataflows
that are not able to hydrate yet. In this commit it is only defined, but
not yet used.