persist: speed up `select count(*) from persist`
Motivation (h/t bkirwi): it's empirically very common for users to do
`select count(*) from x` queries to explore a brand new relation `x`.
These queries run in time proportional to the number of triples in `x`,
which can be slow when `x` is large. However, users prefer when their
queries are fast.

Persist recently started storing in metadata the sum of diffs for each
hollow part in a shard. In `persist_source`, if a part contains data
from entirely before the query `as_of`, and the pushed-down MFP projects
to an empty list of columns, and we can prove that the part is error
free, then we can use the count from stats instead of loading the data.

- Summing the diffs in a part is equivalent to projecting the row in
  each tuple to an empty row and then consolidating. (Which is pretty
  much how select count(*) queries get compiled today.)
- If the as-of timestamp falls in the middle of a part, we can just
  fetch and process the part as normal. The optimization can still
  provide a speedup for other parts.
- This is implemented entirely in persist/`persist_source`; the output
  looks precisely the same to Compute, just a bit more consolidated.

Touches #20618