persist-txn: add dyncfg to apply_eager on init instead of forget_all
In #25992, we discovered that persist-txn forget records do not
consolidate out of the raw shard as intended (they work fine in-mem).
This combined with the forget_all call as part of boot creates a risk
that a crashlooping envd process will quickly increase the physical size
of the txns shard.

Luckily, the forget_all call was only needed as part of the story for
rollback back to the old tables implementation, which has since been
removed. We can now get away with a lesser set of invariants at boot.
Put this behind a dyncfg because it's occasionally useful to have the
forget_all.

Mechanically, I add the dyncfg with a default of the previous behavior
in this commit. In the next, I'll switch the default. This allows us to
potentially cherry-pick just this commit into the currently qualifying
0.92.x release to get the knob out sooner.