Updated Tables
Updated the parameter table to include the AWS Console column as 2 of the fields are different to make the call out more explicit.

`binlog retention hours` requires a MySQL Database connection and SQL to modify the values. Made this its own step.

Moved the table to be below Step 3 as multiple testers blew past the first 2 required steps and jumped to step 3.
Added a validation step