adapter: Removed old builtin objects migration
Sometimes builtin objects in `builtin.rs` are removed, renamed, or
have their type changed (i.e. from table to source). This causes a new
entry in the SystemGidMapping durable catalog collection to be created.
Previously, the old entry was not removed and was ignored by the
catalog. That was a waste of space and caused confusion when reading in
all the entries from disk. This commit removes all the old entries that
are in the durable catalog.

Resolves #26058