mzexplore: add `clone defs` command
Add a `clone defs` command that creates a DDL script that will clone a
subset of the existing schema into a separate namespace and will create
index and view maintenance dataflows in a separate cluster.

This is useful for testing the runtime impact of plan regressions on
clsuters defined with specific optimizer features (see #25273).