storage: fix subsource frontier reconciliation
Storage replicas only report frontiers of collections that are present
in `reported_frontiers`. Entries for subsources are inserted there when
a replica receives the `RunIngestions` command that creates them.
However, the reconciliation logic would only retain the entries of
top-level sources and remove entries of subsources from
`reported_frontiers`, regardless of whether or not these subsources were
still expected. The correct behavior is to keep around the entries for
expected subsources as well.