persist: remove Blob::set atomicity argument
This is an artifact from the single-binary persist days, which only used
Blob, not Consensus. It was introduced so that data blobs could be
written without the file rename trick, and state blobs could be written
atomically. Since the only way to find out about a blob key was through
state, we didn't have to worry about partial writes for data blobs.

In single-binary persist, File was a reasonably supported production
option, so we cared more about the performance. But in platform persist,
we only use s3 for Blob, which is naturally atomic. Additionally, we've
now added things like Blob::list_keys_and_metadata, which have
introduced a way to find blobs that does not come via state.

I don't see us switching to non-atomic writes for data blobs, since it
would add yet another thing to think about and only affect test
performance. So, remove it.