adapter: Make fencing more robust (#26195)
Previously, there were a couple of places in the code where the adapter
should have halted in response to a fence error from the catalog, but
instead was returning a user error. This commit fixes those spots by
halting the adapter in response to a fencing error from the catalog.