storage: move StatsCursor into mz-storage-operators
Motivation is that it's a bit unfortunate to have an mz-storage-types
dep on mz-persist-client. StatsCursor is only used by compute and it
seems reasonable that this would stay true (e.g. `EXPLAIN FILTER
PUSHDOWN` only needs `RelationPartStats`). mz-storage-operators is a bit
of an odd landing place, given that it's not literally an operator, but
we use it to implement operators in compute, so I think it's close
enough. We can always split it out into a new crate later.

Pure refactor, no behavior change.