adapter: acquire read holds before making assumptions about collection sinces
This is preparatory work for #24845, where we factor a
`StorageCollections` out of present-day `StorageController`. And this
`StorageCollections` will be shareable and will come with the
possibility of fontiers advancing concurrently.

We therefore need to guard places where we make assumptions about
frontiers by first acquiring read holds. This will "lock" sinces in
place where they currently are and then look at those sinces.

Before this change, ReadHolds has no protection against being leaked:
when someone forgets calling release_read_holds before dropping a
ReadHolds the holds will be leaked. We need to fix this because with the
above change we're now using ReadHolds in many more places.

We give each ReadHolds a channel on which it sends it's inner
InternalReadHolds when it is dropped. And the Coordinator works off that
channel and releases dropped read holds.

This approach will make tracing/debugging problems slightly harder,
because there is more messaging and more state going around. But I think
we will need something like this for a more concurrent, future
Coordinator where we can acquire read holds, ferry them off to other
tasks and eventually release them there, without mutable access the the
Coordinator.