doc/user: MVP documentation for logical timestamp selection
As part of the mz_now() improvements project [0], we agreed to lightly
improve the documentation of logical timestamp selection. This PR is an
attempt at those improvements.

There's still quite a bit more to explain about how logical timestamp
selection works and its implications for performance, but, with this PR,
at least we have *something* to point users at when they ask about how
the logical timestamp is related to wall clock time.

[0]: https://www.notion.so/materialize/mz_now-braindump-3acde85e02d84cb09808d7c63dab009a#00a22a93afc24c468400bf0e656969ee