bootstrap: fix as-of selection for compute reconciliation
This commit fixes another issue with dataflow as-of bootstrapping that
could disable compute reconciliation. The oversight here was that an
envd restart can make index write frontiers regress (since they are
reset to their newly selected read frontiers), which makes it likely
that using these frontiers to determine an as-of for compute dataflows
will produce an as-of that is earlier than the frontiers of already
installed dataflows have been allowed to compact to.

The solution to this issue is to select as-ofs based on the frontiers of
(transitive) storage dependencies instead. Storage frontiers cannot
regress, so they are guaranteed to provide an as-of that is far enough
advanced for reconciliation to succeed.

The knowledge about transitive storage dependencies was not yet
available during bootstrapping, so this commit adds it by converting the
"index dependent MVs" collection step into a more generic "storage
constraints" collection step that gathers all storage dependencies and
storage dependants needed to constrain the initial timestamp for each
dataflow.

Finally, this commit merges the two `bootstrap_*_as_of` methods into a
single one, to get rid of all the accumulated code and comment
duplication.