bootstrap: fix collection of dependent MVs
Collection of MVs depending on compute indexes had a bug where it uses
`Iterator::any` to determine if any of a number of inserts into a
`BTreeSet` changed the contents of the set. This didn't consider the
fact that `any` is short-circuiting, so if any one insert would return
`true`, subsequent ones would be skipped.

This is fixed here by replacing the use of `any` with a `for` loop.