coord: Fix the new dump endpoint (#26237)
@teskje pointed out that the Coordinator's new `.../dump` endpoint
breaks in practice because a few maps have keys that are not `String`s,
which breaks the JSON serialization. I added a check for this in
testdrive which didn't work because even on errors we would return a
200.

This PR changes the endpoint to return a 500 if we fail to serialize the
state, which should fix the check. It also updates the dump impl to use
the `Debug` format for fields to prevent this method from silently
breaking in the future

### Motivation

Fix the Coordinator's new state dump API

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a

---------

Co-authored-by: Jan Teske <jteske@posteo.net>