adapter: Retry failed table registers
This commit retries failed table registers that fail due to an invalid
upper. Persist can sometimes advance the upper of a table, in response
to some command, without informing the timestamp oracle. For example,
when dropping a table persist txns will advance the upper of all
tables by 1. As a result, the adapter may try and register a new table
with a register timestamp from the timestamp oracle that is invalid
(i.e. too low). Instead of panicking, we should retry the register with
a new valid timestamp.

Works towards resolving #24963
Works towards resolving #25962