adapter: persist RETAIN HISTORY for indexes
Use the same call paths as existing retain history. This ended up cascading into
removing various plan::Index things. It would have been nice to keep all that,
but since this was the only index option, probably reasonable to remove it.

Note that indexes are *parsed* as a AlterIndex, while MVs and tables are parsed
as AlterRetainHistory. Probably we should choose one (Alter<Type> likely), but
keep the planning as is. This can happen later.