storage: Refactor table drop lifecycle
Previously, dropping a table looked something like the following:

  1. Queue a persist command to downgrade the read frontier of the
     table to the empty antichain.
  2. Queue a compaction command to compact the table to the empty
     antichain.
  3. At some later point the storage controller processes all the
     pending compaction commands.
  4. The table shard is forgotten from the persist txn shard.
  5. Once the shard is forgotten and compacted, a command is queued to
     finalize the shard.

This commit updates this logic so that step (4) and (5) are done first.
The logic now looks like the following:

  1. The table shard is forgotten from the persist txn shard.
  2. Once the shard is forgotten, a command is queued to finalize the
     shard.
  3. Queue a persist command to downgrade the read frontier of the
     table to the empty antichain.
  4. Queue a compaction command to compact the table to the empty
     antichain.
  5. At some later point the storage controller processes all the
     pending compaction commands.

Note that we no longer await downgrading the read frontier before
finalizing the shard. Shard finalization is designed to queue shards
that are not fully downgraded and finalize them at a later point. So
this doesn't introduce any correctness issues, but may delay shard
finalization of tables.

This change in logic allows us to make the following improvements:

  - Use a timestamp from the timestamp oracle to forget the table's
    shard from the persist txn shard.
  - Batch the table shard forgets and finalization.

The timestamp improvement fixes a correctness issue where the txn shard
upper would get out of sync from the timestamp oracle. The adapter
makes the assumption that all table uppers are in sync with the
timestamp oracle. Previously, this assumption was broken because
forgets where making up their own timestamps inside the storage
controller, and then advancing the upper of all tables to this
timestamp.

The batch improvement is purely a performance optimization.

Fixes #25962
Fixes #24963