compute-client: schedule transient collections immediately
This ensures that interactive queries (`SELECT`s, `SUBSCRIBE`s) are
always responded to as quickly as possible. Delaying scheduling until
all inputs are ready can delay responses to such queries by the time it
takes for frontier updates to travel from the replica to the controller
and, worse, by the time until the maintenance ticker fires again, which
shows up in our various nightly benchmarks.