adapter: rename arrangement merge effort option
This commit renames the cluster/replica option `IDLE ARRANGEMENT MERGE
EFFORT` to just `ARRANGEMENT MERGE EFFORT`.

Dropping the "idle" removes a potentatial source of confusion as the
configured merge effort isn't actually applied only when a worker is
idle, but on every worker step.

This serves also as preparation for making that option control the new
proportionality-based arrangement merging strategy, rather than DD's
`idle_merge_effort` option. The new name is generic enough that it can
apply to any merging strategy we might want to use in the future.