test: add `index-compute-dependencies` restart test
Various parts of the optimizer internals and tooling, such as

- `EXPLAIN REPLAN`
- `bin/mzcompose clone defs`

are currently depending on the fact that the `GlobalId` ordering
respects dependency ordering. In other words, if an index `i` is created
after a catalog item `x`, then `x` cannot use `i` even after restart.

Add a test to codify this assumption so we can get an early signal if
this is broken for some reason in the future. To execute the test run:

```bash
bin/mzcompose --find restart run index-compute-dependencies
```