adapter: remove `IDLE ARRANGEMENT MERGE EFFORT` option
This commit removes the `IDLE ARRANGEMENT MERGE EFFORT` cluster/replica
option. The option has always been documented as "unstable" and we don't
know of anyone relying on it in production.

Note that we don't introduce a replacement option. The assumption is
that none is needed as we either have no need to change the arrangement
merge effort in practice, or having the feature flag should suffice.
It's not worth keeping around a significant amount of code just for the
very small chance that it might be useful.