catalog: Add durable pre-open migrations
Previously, if someone wanted to write a catalog migration they had two
options:

  1. Increment the catalog version and write a catalog schema/data
     migration. This process is not light-weight and involves working
     directly with protobuf.
  2. Write a catalog migration after all the catalog data has been
     loaded into memory.

If you wanted to, for example, modify all clusters, then you would be
forced to write a schema/data migration (option 1). Catalog migrations
(option 2) happen after the clusters have already been loaded and
would not work. schema/data migrations can be annoying to write and are
over-kill for this situation since the schema of catalog objects are
not changing.

This commit adds a third type of migration that allows modifications
to the durable catalog before it's been loaded into memory.