persist-txn: Use efficient timestamp rewrites (#26316)
Previously, if persist-txn failed to commit a transaction, then it
would delete the batch associated with the failed transaction and
create a new batch for every commit retry. Deleting and creating
batches are potentially expensive operations.

In 1f861336ce368fffecfdb11c8277816243257d4a, an efficient method was
added to rewrite the timestamp of an existing batch. This commit uses
that new method in persist-txn to rewrite the timestamp of a batch when
a commit fails instead of deleting and creating a new batch.

Part of #22173