adapter: Provide linearizable boot timestamp
Previously, we would provide a non-linearizable boot timestamp to
`Catalog::open`. A non-linearizable timestamp is not written down, and
therefore it's possible that a later environment boots up with a lower
timestamp. We were unable to provide a linearizable timestamp due to a
bootstrapping problem. We needed to look in the catalog to see what
timestamp oracle implementation to use, but we needed to know which
timestamp oracle implementation to use to get a linearizable timestamp
to open the catalog.

Now that there's only a single timestamp oracle implementation, this
bootstrapping problem has gone away. This commit updates the bootstrap
process so that a linearizable timestamp is passed to `Catalog::open`.
This is useful for things like adding audit log events in a migration,
which require linearizable timestamps.