sql: Set disk_cluster_replicas_default sys var
sqllogictest requires that the disk_cluster_replicas_default system
variable is set to true, or it will fail. This can be confusing on not
obvious if someone is running sqllogictest outside mzcompose.This
commit automatically sets this variable to true if it's unset and
panics if it's set to the wrong value. Additionally, it adds some
structure to add more required default system variables in the future.