adapter: also delay view types
This fixes key discrepancy problems because `HirRelationExpr::typ` doesn't set
keys, while `Mir` does.

But because views depend on each other, some must wait for the typ to be
available, which means we need something to do the waiting. Have a very bad
thread sleep retry loop. This does mean that we serialize work, and lose a bit
over 1s from the previous startup time.