catalog: Refactor durable code visibility
Previously, all usages of the DurableCatalogState and
OpenableDurableCatalogState used dynamic dispatch. However, the APIs in
`mz_catalog::durable` returned a concrete struct on the stack and
required callers to box the struct. This resulted in
`mz_catalog::durable` making a lot of unnecessary implementation
details public to outside crates.

This commit updates the `mz_catalog::durable` APIs to directly return
boxed trait objects, so that the implementation details can be hidden
from outside crates.