catalog: Add infrastructure for listening catalog (#26359)
This commit adds the initial infrastructure for the in-memory catalog to
update itself in reaction to changes made by the durable catalog.
Additionally, it implements the update logic for databases. Future
commits will add the update logic for all catalog objects.

The update logic is used when opening the catalog to initialize the
in-memory catalog. A more obvious place to use the update logic might
have been during catalog transactions. However, currently catalog
transactions update both the in-memory catalog state and a durable
transaction for each operation. The end goal is to have the catalog
transaction only update a durable transaction and then update the
in-memory state using the update logic after the durable transaction
commits. Using the update logic in a catalog transaction before the
update logic is complete for all objects means that for some objects
we'd have to look in the durable transaction for the current state,
while for other objects we'd have to look in the in-memory catalog for
the current state. Getting this right would have been extremely hard and
error-prone. Once the update logic is complete for all objects, we'll
start using it in catalog transactions.

Works towards resolving #24844