sql: implement ALTER RETAIN HISTORY (#26254)
Support altering and resetting the retain history objects that can be created with it. This is a bit messy because it works in three different ways depending on the statement type. It's either a new parsed statement type, or a plan type, or needs special detection during execution (sources, because retain history shouldn't be purified). We are ok with this divergence for now because this is a new thing and we'd like to get some feedback about it before committing more time into unifying the implementations.

### Motivation

  * This PR adds a known-desirable feature.

### Tips for reviewer

* Although this will be squashed, viewing each commit may ease the
review work.

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly
considered.
- [x] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [x] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [x] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - n/a

---------

Co-authored-by: Rainer Niedermayr <rainer@materialize.com>