adapter: initialize compute/storage "read capabilities" when needed
Before, we were first creating all the compute/storage objects in one
batch and _then_ we were initializing all the read
capabilities/policies. This will not work when objects within the batch
want to acquire read holds on other objects in the batch, which we are
doing in the next commit.

Now, we ensure that a read capability is there when needed, and patch it
up with the correct read policy when we know it.