storage/pg: stream large transactions to persist
This PR makes a slight modification to the capability management logic
of the postgres source such that the capability that governs the minting
of reclock bindings is advanced eagerly, before a transaction is
ingested. This has the effect that the data of the transaction will not
accumulate in the reclock operator and will instead be passed
downstream, where the persist sink operator can stream them to S3
batches.

Fixes #26257

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>