adapter: create all outstanding collections at once
The adapter previously called create_collections once per source
when creating sources. However, the new subsource structure is
more amenable to just creating all sources at once.

For instance, each call to create_collection that creates a
subsource modifies the source's definition and requests that it's
rescheduled. Doing this once for each subsource is wasteful when
we know that before we will let the source run at a steady state
we plan to repeatedly interrupt it.

Note that this is only a performance optimization and does not
have any semantic implications.