persist: introduce proc macro to run unit tests with various configs
We've gotten somewhat lucky so far to be able to more-or-less pick a
default set of configuration values for persist unit tests, perhaps
adding a couple new tests for the non-default cases. But inline writes
probably wants to keep testing coverage of both the enabled and disabled
path in perpetuity.

So, introduce a new `mz_persist_proc::test` prod macro that runs an
entire cargo test under an arbitrary set of configurations and hook it
up to persist's tests.

Intentionally don't add a lint requiring its usage because many tests
don't use configuration and so running them multiple times would be pure
waste. Particularly egregious would be the various randomized prop
tests (e.g. for encoding).