compute-client: introduce `ReadCapability` response
This commit introduces a `ReadCapability` compute response that will
have the purpose of reporting which input times still need to be kept
available for a collection, and therefore which input read holds the
controller needs to ensure. Previously this information was inferred
from `FrontierUpper` responses, but with the introduction of `REFRESH`
materialized views that can jump their upper frontier ahead of their
input readers, this approach has stopped working.

Here the new response type is only defined, but it is not used yet.