compute: early consolidation in `persist_sink`
This commit removes the `CorrectionBuffer` type from the compute
`persist_sink` and instead makes `persist_sink` keep all pending updates
in the `UpdateStash`. This makes it possible to continuously consolidate
`desired` and `persist` updates as they arrive, rather than only doing
so once a new batch is about to be emitted.

The hope is that this change will significantly reduce `persist_sink`'s
hydration memory spike.