tests: sleep in scalability tests to allow dataflows to be fully hydrated
There is an interaction between delayed dataflow scheduling (on the
compute side) and how the new StorageCollections learns about the uppers
of collections that makes it so we slightly delay scheduling of
dataflows in cases. This would show up in benchmark results as the very
first select (or similar) taking about a second.