catalog: Add more structure to state init writes
Previously, initializing the in-memory catalog state involved a mixture
of reading and writing from and to the durable catalog state. This
commit adds some more structure to the writes involved in this process.
The writes are not removed completely, but they are grouped into three
distinct groups instead of mixed throughout the process. Those groups
are:

  1. Before any data is read into the in-memory catalog state, new
     builtin objects are written down durable and removed builtin
     objects are removed from the durable state. Currently, we don't
     check all types of builtin objects for removal, and it's likely
     that certain types are leaked in the durable state. This stage
     can/should arguably be done at a lower level and should be
     completely handled by the durable catalog without any involvement
     from the in-memory catalog. That refactor may be done in a later
     commit. Durably recording system variables from LaunchDarkly
     piggybacks off this stage and is included out of convenience.
  2. After all data is read into the in-memory catalog state, user item
     ASTs are migrated, if any migrations are defined.
  3. After user item ASTs are migrated, all builtin items whose schema
     has changed since the last version are migrated.

This structure subjectively makes the state initialization easier to
reason about and more importantly will make parts of the
multi-subscriber catalog easier to write.

Works towards resolving #24844