dyncfg: allow a given config type to have multiple default types
Currently, for every type T that implements `ConfigType`, there is
exactly one type, `T::Default`, that can be used as the default value
for that configuration type. For example, the default type for a `bool`
configuration value is `bool`. The interesting case is `String`, for
which the default type is `&'static str`; this allows instantiating
`Config` structs in constants, where `Strings` cannot be allocated.

This commit allows each T to be constructed from any number of default
types via the introduction of a new `IntoConfigType` trait. There are
two blanket implementations of `IntoConfigType`:

  * An implementation for `T`, which permits every config type to be
    used as its own default. This is the common case, and supports
    primitive config types like `bool` that do not require allocation.

  * An implementation for `fn() -> T`, which permits using a callable as
    the default value. This allows for flexibility when initializing
    complex types like strings and, in a future commit, JSON values,
    where the default value may require a heap allocation.

    The persist untrimmable stats configuration is adjusted to use a
    callable default as a demonstration of its flexibility.

Finally, there is one concrete implementation of `IntoConfigType` to
retain the special case of allowing `String` configuration variables to
be initialized from a `&'static str` value is