adapter: clearer distinction between TimelineReadHolds and ephemeral ReadHolds
Preparatory work for
https://github.com/MaterializeInc/materialize/issues/24845, where we
want to introduce more concurrency to the Coordinator and Controllers.

The code/comments/docs explain the motivation and how/why they are used.
The overarching motivation is that future work around introducing more
concurrency in the adapter/Coordinator will make this necessary, but it
also feels correct to make this distinction that has always been there
explicit in the code and interfaces. A previous refactoring was not
going far enough.

It's nice to see how the previous
acquire_read_holds/update_read_holds/release_read_holds were actually
used for disjoint purposes, we don't have to duplicate those methods
now.

Also, this moves initialize_read_holds inside the methods where it's
used because its semantics were confusing and because it was really only
used in two places: for the thing we now call TimelineReadHolds and what
we now call ephemeral ReadHolds.