adapter: use STORAGE ReadHold instead of custom hold tracking in adapter
Preparatory work for #24845, where we want to introduce more concurrency
to the Coordinator and Controllers.

Before, adapter would track outstanding `ReadHolds` in its own state and
continually update the `ReadPolicy` of storage collections based on a
base policy and outstanding read holds.

Now, we go directly to STORAGE for acquiring ephemeral `ReadHolds`. Note
that the adapter still keeps `TimelineReadHolds` for collections, which
control the `ReadPolicy` that we continually update. This is so that
timeline-dependent collections remain readable at the oracle read
timestamp even when there are no outstanding other (ephemeral) read
holds.

This leaves the COMPUTE part of `ReadHolds` untouched.