adapter,sql: add EXPLAIN PLAN INSIGHTS statement
EXPLAIN PLAN INSIGHTS is a new feature that produces key insights about
the plan for a SELECT statement.

In its initial form, its only consumer will be the console, which will
use it to power hints about e.g. missing indexes
(MaterializeInc/console#1977). This allows us to keep the initial
implementation simple and focused--e.g., only the JSON format is
supported for now.

In the future, if the statement proves useful, we can consider
documenting the statement for end user use, and introducing a
human-readable textual format.

There's no feature flag for the statement, as the EXPLAIN statement is
not subject to backwards compatibility guarantees.

Work towards #26436.