compute: produce `Export` events for logging dataflows
This commit arranges for `Export` events to be logged for logging
dataflows. Previously, these events were only logged for dataflows
created through `handle_create_dataflow`, which I think was an
oversight. At least various places in the logic of the compute logging
dataflow assume that an `Export` event is emitted for each dataflow
before any other events are emitted.