compute: `CollectionLogging` type
This commit introduces a `CollectionLogging` type that encapsulate
logging of `ComputeEvent`s for a given compute collection. The type
remembers previously logged events and ensures they are correctly
retracted when relevant properties of the collection are updated or the
collection is dropped.

For now `CollectionLogging` only supports `Export` and `Frontier`
events. A second commit will add support for `ImportFrontier` events as
well.