*: remove `mz_compute_delays_histogram`
This introspection relation has proven to be of limited use, given that
we now have `mz_materialization_lag` which fulfills a similar purpose
but is more generally usable. Its contents are also hard to interpret
because (a) the reported delays are overestimates and (b) the reported
delays will be zero most of the time for `REFRESH` materialized views.

`mz_compute_delays_histogram` does thus not provide enough value to
justify its considerable implementation complexity, so this commit
removes it. Not only does this let us delete a bunch of code, but it
also enables changing the way we log import frontiers, which will
simplify the planned work of adding `ReadCapability` compute responses.